+++ playwright/tests/admin/homepage.spec.js
@@ -5,7 +5,7 @@
  */
 
 // @ts-check
-import { test, expect } from '../utils/adminTest';
+import { test, expect } from '../utils/admin/adminTest';
 
 test('admin home page', async ({ page, adminPage }) => {
   await adminPage.signInAsNewAdminUser();
