+++ playwright/tests/utils/admin/adminTest.js
@@ -6,7 +6,8 @@
 // @ts-check
 import * as base from '@playwright/test';
 
-import { BaseTestPage } from './baseTest';
+import { BaseTestPage } from '../baseTest';
+
 import WaitForAdminPage from './waitForAdminPage';
 
 /**
